refactor: centralize errors in api
